afs: Allow afs_GetUser to return errors
Change afs_GetUser to allow for returning error codes, to let future
changes make it so afs_GetUser can fail. Change all callers to
accommodate errors, though this commit itself does not introduce any
possible errors.

For afs_Conn() and afs_ConnByHost(), set areq->volumeError if we get an
error from afs_GetUser(). If we don't set areq->volumeError, then
afs_Analyze() will treat a NULL connection as an indication of a network
problem, which is not accurate. That can cause afs_Analyze() to do a
hardmount retry loop (if hardmount is enabled), and otherwise treat the
error as a network error.

To distinguish this error from the cases where the requested volume
doesn't exist or is busy, set areq->volumeError to a new constant,
VOL_INTERNAL_ERROR. This is a bit awkward, since the error isn't
actually related to the volume, but treating the error similarly to a
nonexistent volume is the behavior we want: we don't want to retry the
request, we should just fail with an internal error immediately.

Change-Id: Idf05cae1d2db5c2a20933dc69056f2988fd694f7
Reviewed-on: https://gerrit.openafs.org/14088
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>